Closed #2
I added a field that contains a boolean wether the specified url is deleted at autoscout.
The entry stays in the db.
To achieve this I simply double-checked all already visited urls (that are not marked 'deleted').
If the url is still valid it gets saved again in the db with a different timestamp and possible different values.